refactor: Exception -> JwtStatus 반환 로직 공통화
